+++ philo/Makefile
@@ -1,5 +1,6 @@
 CC = gcc
 CFLAGS = -Wall -Werror -Wextra
+CSANITISE = -g -fsanitize=address
 PTHREAD = -pthread
 
 SRCS = $(wildcard *.c)
@@ -11,10 +12,10 @@ NAME = philo
 all: $(NAME)
 
 $(NAME): $(OBJS)
-	$(CC) $(CFLAGS) $(PTHREAD) -o $@ $^
+	$(CC) $(CFLAGS) $(CSANITISE) $(PTHREAD) -o $@ $^
 
 %.o: %.c
-	$(CC) $(CFLAGS) $(PTHREAD) -c $< -o $@
+	$(CC) $(CFLAGS) $(CSANITISE) $(PTHREAD) -c $< -o $@
 
 clean: 
 	rm -f $(OBJS)
